% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_classname}
\alias{alt_classname}
\title{Gets the name of the altrep class of an object}
\usage{
alt_classname(x)
}
\arguments{
\item{x}{An ALTREP object of any class}
}
\value{
The class name as a character scalar.
}
\description{
This is a a human-readable name for the altrep class.
}
\examples{
alt_classname(1:3)
}
