% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_data1}
\alias{alt_data1}
\title{Gets the first altrep data slot}
\usage{
alt_data1(x)
}
\value{
Possibly any R object, including \code{NULL}
}
\description{
ALTREP objects have two data slots, both of which can hold any R type.
The exact meaning of each slot can depend entirely on the ALTREP class and
how it wants to represent data.
}
\details{
Although the exact meaning of each slot is flexible, a \emph{convention} used in
R core is
for \code{data1} to hold the "compressed" state of a type, and for \code{data2} to
hold the "expanded" state. See the \code{compact_seq} vignette for more information
}
\examples{
alt_data1(1:3)
}
