% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_data2}
\alias{alt_data2}
\title{Gets the second altrep data slot.}
\usage{
alt_data2(x)
}
\arguments{
\item{x}{An ALTREP object of any class}
}
\value{
Possibly any R object, including \code{NULL}
}
\description{
ALTREP objects have two data slots, both of which can hold any R type.
The exact meaning of each slot can depend entirely on the ALTREP class and
how it wants to represent data.
}
\note{
\strong{Warning}: storing the result will cause an R session crash with a
\code{deferred_string} ALTREP object.
}
\examples{
alt_data2(1:3)
}
