% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_details}
\alias{alt_details}
\title{Returns a summary of the properties of an ALTREP object}
\usage{
alt_details(x)
}
\arguments{
\item{x}{An ALTREP object of any class}
}
\value{
A list with the following entries:
\itemize{
\item \code{class_name} (character scalar): see \code{\link[=alt_classname]{alt_classname()}}
\item \code{pkg_name} (character scalar): see \code{\link[=alt_classname]{alt_classname()}}
\item \code{base_type} (character scalar): see \code{\link[=alt_type]{alt_type()}}
\item \code{data1} (anything): see \code{\link[=alt_data1]{alt_data1()}}
\item \code{data2} (anything): see \code{\link[=alt_data2]{alt_data2()}}
}
}
\description{
Returns a summary of the properties of an ALTREP object
}
\note{
\strong{Warning}: storing the result will cause an R session crash with a
\code{deferred_string} ALTREP object.
}
\examples{
alt_details(1:4)
}
