% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{alt_inspect}
\alias{alt_inspect}
\title{Inspects an R data structure}
\usage{
alt_inspect(x)
}
\arguments{
\item{x}{Any R object to inspect}
}
\value{
\code{x}, invisibly
}
\description{
Specifically this prints a kind of tree that includes memory addresses,
\code{SEXPTYPE}s, and a few ALTREP tidbits.
This is useful in particular because ALTREP classes tend to have custom
implementations for inspect, which can be a useful way to summarise
ALTREP state.
}
\note{
This is a re-export of an internal R function, so outputs shouldn't be
assumed to be stable.
}
\examples{
alt_inspect(1:5)
}
