% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_touch_dataptr}
\alias{alt_touch_dataptr}
\title{Touches the DATAPTR of a vector}
\usage{
alt_touch_dataptr(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
\code{x}, invisibly, unchanged.
}
\description{
This is a utility function that accesses the \code{DATAPTR} of a vector, but
doesn't touch it or do anything else. The \code{DATAPTR} is a pointer to the
actual array data of the vector. This is useful because it often triggers
certain behaviours in ALTREP vectors like expanding them into full form.
}
\examples{
x <- 1:5
compact_is_expanded(x)
alt_touch_dataptr(x)
compact_is_expanded(x)
}
\keyword{advanced}
