% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_type}
\alias{alt_type}
\title{Gets the name of the type that this ALTREP is representing.}
\usage{
alt_type(x)
}
\arguments{
\item{x}{An ALTREP object of any class}
}
\value{
The name of the fundamental vector type of that \code{x} is representing,
as a \link{character} scalar. For example "integer" or "character".
}
\description{
This will almost certainly return the same result as \code{typeof(x)}. However,
the author would be interested to know if it doesn't!
}
\examples{
alt_type(1:3)
}
