% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compact_details}
\alias{compact_details}
\title{Returns a list containing the fields of a compact vector}
\usage{
compact_details(x)
}
\arguments{
\item{x}{An ALTREP vector of class \code{compact_realseq} or \code{compact_intseq}}
}
\value{
A list with the elements:
\itemize{
\item \code{length} (integer scalar): the length of the sequence
\item \code{start} (integer scalar): the first value of the sequence
\item \code{step} (integer scalar) : the increment between entries in the sequence
\item \code{expanded} (double, integer vector, or \code{NULL}): a non-ALTREP version of
\code{x}, or \code{NULL} if it hasn't been expanded
}
}
\description{
Returns a list containing the fields of a compact vector
}
\examples{
compact_details(1:5)
}
