% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compact_expand}
\alias{compact_expand}
\title{Forces a compact vector into extended form}
\usage{
compact_expand(x)
}
\description{
This will allocate the entire vector as a standard representation vector,
and store it in the \code{data2} slot.
}
\examples{
x = 1:3
compact_expand(x)
alt_inspect(x)
}
