% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compact_to_standard}
\alias{compact_to_standard}
\title{Returns a non-ALTREP copy of a compact vector}
\usage{
compact_to_standard(x)
}
\arguments{
\item{x}{An ALTREP vector of class \code{compact_realseq} or \code{compact_intseq}}
}
\description{
Returns a non-ALTREP copy of a compact vector
}
\examples{
compact_to_standard(1:5)
}
