% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deferred_expand}
\alias{deferred_expand}
\title{Forces a deferred string into extended form}
\usage{
deferred_expand(x)
}
\description{
Forces a deferred string into extended form
}
\examples{
x = as.character(1:3)
deferred_expand(x)
alt_inspect(x)
}
