% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_altrep}
\alias{is_altrep}
\title{Checks if an R object is ALTREP}
\usage{
is_altrep(x)
}
\value{
A scalar logical
}
\description{
This checks if \code{x} is an instance of an ALTREP class. Notably it doesn't
check if \code{x} \strong{is} an ALTREP class, which is more difficult to achieve.
}
\examples{
is_altrep(1)
is_altrep(1:2)
}
