% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_deferred_string}
\alias{is_deferred_string}
\title{Checks for deferred string ALTREPs}
\usage{
is_deferred_string(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
Logical scalar. \code{TRUE} if \code{x} is a deferred_string ALTREP,
otherwise \code{FALSE}
}
\description{
Checks for deferred string ALTREPs
}
\examples{
is_deferred_string("hi")
is_deferred_string(as.character(1))
}
