% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_mmap}
\alias{is_mmap}
\title{Checks for memmap vector ALTREPs}
\usage{
is_mmap(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
A scalar logical. \code{TRUE} if \code{x} is a mmap vector, otherwise
\code{FALSE}
}
\description{
Checks for memmap vector ALTREPs
}
\examples{
is_mmap(1)
mmap_make(data = 1:100) |> is_mmap()
}
