% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_wrapper}
\alias{is_wrapper}
\title{Checks for wrapper object ALTREPs}
\usage{
is_wrapper(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
Logical scalar. \code{TRUE} if \code{x} is a wrapper ALTREP,
otherwise \code{FALSE}
}
\description{
Checks for wrapper object ALTREPs
}
\examples{
is_wrapper("5:1")
sort(5:1) |> is_wrapper()
}
