% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmap_details}
\alias{mmap_details}
\title{Returns a list containing the fields of a memory mapped vector}
\usage{
mmap_details(x)
}
\arguments{
\item{x}{A vector which has an mmap ALTREP class}
}
\value{
A list with the fields:
\itemize{
\item \code{file_name}: A character scalar. The file name of that file that was
mapped
\item \code{size_bytes}: An integer scalar. The number of bytes of the memory
map
\item \code{length}: An integer scalar. The number of entries in the memory mapped
vector
\item \code{type}: "double", or "integer", indicating the data type of the memory
map
\item \code{ptrOK}: logical scalar. Unknown meaning
\item \code{wrtOK}: logical scalar. \code{TRUE} if writing to the memory map is allowed
\item \code{serOK}: logical scalar. Unknown meaning
}
}
\description{
Returns a list containing the fields of a memory mapped vector
}
\examples{
mmap_make(data = 1:10) |> mmap_details()
}
