% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memmap.R
\name{mmap_make}
\alias{mmap_make}
\title{Makes a memory-mapped vector}
\usage{
mmap_make(path = NULL, data = NULL, type = NULL)
}
\arguments{
\item{path}{Path to an existing file, or a file to write to if \code{data} is
provided}

\item{data}{A double or integer vector to write to \code{path}, or \code{NULL} to
only read from it.}

\item{type}{Either "double" or "integer", indicating the data type to
read from disk. If not provided, it will be assumed to be the type of
\code{data}.}
}
\value{
A memory mapped vector, using the contents of \code{path}
}
\description{
Makes a memory-mapped vector
}
\examples{
mmap_make(
  tempfile(),
  rnorm(n=5)
)
}
