% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_alt_data1}
\alias{set_alt_data1}
\title{Sets the \code{data1} value of an ALTREP}
\usage{
set_alt_data1(x, value)
}
\arguments{
\item{x}{An ALTREP object to modify}

\item{value}{The new value of the \code{altrep_data1} slot}
}
\value{
\code{x}, invisibly (because \code{x} has been modified in-place you generally
won't want or need to store the return value)
}
\description{
\strong{Don't use this function unless you know what you're doing} or just
don't care about your R session.
If you set the data to some
unexpected value you are very likely to cause a SEGFAULT and crash the
entire R session. Also, this modifies the existing object in-place, meaning
that all variables pointing to this same ALTREP will be modified.
}
\examples{
x <- 1:5
set_alt_data1(x, c(10, 10, -1))
x
}
\keyword{advanced}
