% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wrapper_details}
\alias{wrapper_details}
\title{Returns a list containing the fields of a wrapper object}
\usage{
wrapper_details(x)
}
\arguments{
\item{x}{Any R object belonging with a \code{wrap} ALTREP class}
}
\value{
A list with the following entries:
\itemize{
\item \code{contents}: Any R object. The underlying object being wrapped
\item has_na\verb{: Logical scalar. }TRUE\code{if}contents\verb{**might** contain}NA\verb{, or }FALSE` if it \strong{definitely doesn't}.
\item \code{is_sorted}: Logical scalar. \code{TRUE} if \code{contents} is sorted in any
order
\item \code{descending}: Logical scalar. \code{TRUE} if \code{contents} is sorted in
descending order, \code{FALSE} if it is sorted in ascending order, or
\code{NA} if it isn't sorted or its sorting status is unknown.
\item \code{na_first} Logical scalar. \code{TRUE} if \code{contents} is sorted in an order
that puts \code{NA} values first, \code{FALSE} if it is sorted last, or \code{NA}
if \code{contents} is unsorted or its sorting status is unknown
}
}
\description{
Returns a list containing the fields of a wrapper object
}
\examples{
sort(5:1) |> wrapper_details()
}
