% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper_make}
\alias{wrapper_make}
\title{Makes a custom wrapper ALTREP}
\usage{
wrapper_make(
  x,
  has_na = FALSE,
  is_sorted = NA,
  descending = FALSE,
  na_first = FALSE
)
}
\arguments{
\item{x}{Any R object to wrap}

\item{has_na}{Logical scalar. \code{FALSE} if \code{x} should be considered to be free
of \code{NA} values, otherwise should be set to \code{TRUE}}

\item{is_sorted}{Logical scalar. \code{TRUE} if \code{x} should be considered sorted,
\code{FALSE} if it should be considered unsorted, or \code{NA} if its sorted status
is unknown}

\item{descending}{Logical scalar. \code{TRUE} if \code{x} should be considered to be
sorted in descending order, \code{FALSE} for ascending order.
If \code{x} is unsorted this is unused.}

\item{na_first}{Logical scalar. \code{TRUE} if \code{x} should be treated as if it
has \code{NA} values at the start of the vector, \code{FALSE} if they are at the end.
If \code{x} is unsorted this is unused.}
}
\value{
An object identical to \code{x} but with some hidden ALTREP metadata
applied to it.
}
\description{
Makes a custom wrapper ALTREP
}
\examples{
wrapper <- wrapper_make(c(5, 3, NA, 1), is_sorted = TRUE, has_na = FALSE)
wrapper_details(wrapper)
}
