% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alt_class}
\alias{alt_class}
\title{Gets the altrep class of an object.}
\usage{
alt_class(x)
}
\value{
A raw vector uniquely identifying the altrep class of \code{x}, or \code{NULL}
if \code{x} is not an altrep object.
}
\description{
The altrep class is not the same as the name of the class. For that,
you can use \code{\link[=alt_classname]{alt_classname()}}. The altrep class is a \link{raw} vector whose
value uniquely distinguishes the class from other classes. You are unlikely
to find much meaning in the actual vector values, but you might find it
useful to check if two objects belong to the same class.The other
meaningful use of the class is to check its \code{\link[=attributes]{attributes()}}, but it is
recommended that you instead use the three dedicated wrapper functions for
these attributes: \code{\link[=alt_classname]{alt_classname()}}, \code{\link[=alt_pkgname]{alt_pkgname()}}, and \code{\link[=alt_type]{alt_type()}}.
}
\seealso{
alt_classname alt_pkgname alt_type
}
