% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compact_expand}
\alias{compact_expand}
\title{Forces a compact vector into extended form}
\usage{
compact_expand(x)
}
\arguments{
\item{x}{An ALTREP vector of class \code{compact_realseq} or \code{compact_intseq}}
}
\value{
\code{x}, \strong{not a copy of \code{x}}
}
\description{
This will allocate the entire vector as a standard representation vector,
and store it in the \code{data2} slot.
Note: this modifies \code{x} in-place, so will modify any copies of \code{x}, and
is irreversible.
}
\examples{
x = 1:3
compact_expand(x)
compact_details(x)
}
