% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compact_expand}
\alias{compact_expand}
\title{Expands a compact vector, setting the value of its \code{data2} to a standard
representation vector}
\usage{
compact_expand(x)
}
\description{
Expands a compact vector, setting the value of its \code{data2} to a standard
representation vector
}
\examples{
x = 1:3
compact_expand(x)
alt_inspect(x)
}
