% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deferred_expand}
\alias{deferred_expand}
\title{Forces a deferred string into extended form}
\usage{
deferred_expand(x)
}
\arguments{
\item{x}{Character vector with the \code{deferred_string} ALTREP class}
}
\value{
\code{x}, not a copy of \code{x}, after expansion.
}
\description{
Forces a deferred string into extended form
}
\examples{
x = as.character(1:3)
deferred_expand(x)
alt_inspect(x)
}
